# Conflicts:
#	src/main/java/com/pse/testserver/repository/GroupRepository.java
#	src/main/java/com/pse/testserver/repository/MessageRepository.java
#	src/main/java/com/pse/testserver/repository/PostRepository.java
#	src/main/java/com/pse/testserver/repository/SubcategoryRepository.java

+++ src/main/java/com/pse/testserver/repository/MessageRepository.java
@@ -5,13 +5,8 @@ import com.pse.testserver.entities.User;
 
 import org.springframework.data.repository.CrudRepository;
 
-/*
-
- */
 public interface MessageRepository extends CrudRepository<Message, Integer> {
-    Message findById(int id);
-
+    Message findById(long id);
     Message findBySender(User user);
-
     Message findByReceiver(User user);
 }
